Import relevant firebase materials
