+++ package.json
@@ -17,7 +17,9 @@
     "framer-motion": "^10.12.18",
     "maath": "^0.7.0",
     "react": "^18.2.0",
+    "react-cookie": "^6.1.1",
     "react-dom": "^18.2.0",
+    "react-firebase-hooks": "^5.1.1",
     "react-router-dom": "^6.16.0",
     "react-tilt": "^1.0.2",
     "react-type-animation": "^3.1.0",
