--- src/firebase_setup/firebase.js
@@ -1,24 +1,0 @@
-// Import the functions you need from the SDKs you need
-import { initializeApp } from "firebase/app";
-import { getAnalytics } from "firebase/analytics";
-import { getFirestore } from "@firebase/firestore"
-// TODO: Add SDKs for Firebase products that you want to use
-// https://firebase.google.com/docs/web/setup#available-libraries
-
-// Your web app's Firebase configuration
-// For Firebase JS SDK v7.20.0 and later, measurementId is optional
-// TODO - need a safer way to store the informations here: https://www.makeuseof.com/react-api-keys-store-access/
-const firebaseConfig = {
-  apiKey: "AIzaSyBzRAQygXp7iL6oepOig6sTxubfTHW7nOc",
-  authDomain: "ai-remember.firebaseapp.com",
-  projectId: "ai-remember",
-  storageBucket: "ai-remember.appspot.com",
-  messagingSenderId: "60106174976",
-  appId: "1:60106174976:web:af79f5c1819fff32e65ee5",
-  measurementId: "G-P3ZGZ14ZR6"
-};
-
-// Initialize Firebase
-const app = initializeApp(firebaseConfig);
-const analytics = getAnalytics(app);
-export const firestore = getFirestore(app)
